Move checkForResponse to utils and add it to the login action
